Update mmgbsa_new.py
The problem, turns out, wasn’t GPU related or summit—a sneaky variable re-assignment which called a constructor before the assignees’ deconstructor. The pipeline had three simulations phases to run. Under the hood, platform.getPlatform(“CUDA”) wanted to create cuCtxCreate on a GPU that no one else was using. When that for loop got around to the next phase, platform.getPlatform was called without releasing the prior platoform, so if you wanted device 0, you’d get device one. On JSrun when we would give a resource group only a single gpu, this would result in a failure as there is not another gpu to use, and the only gpu was not released correctly.